Split tiddler text out into separate tables
This is in preparation for adding binary tiddler support. Basically,
binary tiddlers go in the "file" property during POSTs/PUTs. Presence of
the "file" attribute means the tiddler will be a binary tiddler and will
be treated as such (e.g. it will be served in binary/raw form unless
explicitly requested otherwise).

This is handled using a polymorphic "textable" that picks between
TextRevision and FileRevision (yet to be implemented) at revision
creation time.