Add POST support to revisions route
The POST supoprt lets you post a revision id and have it revert to that
revision. This is needed for, among other things, reverting binary
tiddlers. It also paves the way for proper PATCH, etc support.

This commit also introduces the type=file query param to the edit
tiddler form and updates the revert form to use the new POST path.