Add a custom tiddler responder
TiddlerResponder handles sending binary tiddlers in either their
specified content type, or in their html representation.

This decision is made based on whether the htlm format was requested
explicitly (using :format) and whether there is a */* in the Accept
header.

Currently non FileRevision tiddlers always render in html, so this isn't
perfect quite yet as you can't request, say, text/plain in it's raw
format yet unless you upload it as a file.