Rework scopes, relationships and filtering; Add filtering to revisions
This changes the defautl scopes for relationships, introducing a new
relationship in Tiddler that computes the latest revision for each
tiddler directly in SQL. This simplies things and lets us, for e.g.
select tiddlers with a particular tag in the current revision, rather
than tiddlers that have previously had a tag at some point in the past.

This brings with it benefits and drawbacks and I'm not convinced that
it's the right thing to do. The benefits are clearly that it lets us do
filtering in SQL directly rather than  in ruby after fetching the sql
records. This simplifies things a lot. The drawbacks are that it makes
the relationships themselves more complicated (speed isn't currently a
concern) and requires postgres specific SQL statements involving window
functions. This makes migrating to somethig else in the future a lot
harder.

For reference, the sensible-ish alternative is to denormalize slightly
and write everything twice: once for the tiddler, and once for the
revision.