Add linking extensions to markdown
See lib/federacy_markdown.rb for a list of what you can do.

This also introduces a short link syntax for linking to tiddlers and
spaces by their respective names. Note that given that names are not
unique and sybject to change, this is not guaranteed to always link to
correct thing. Short url syntax is:

 - /spaces/:space_id/t/:tiddler_title
 - /s/:space_name
 - /s/:space_name/:tiddler_title
 - /u/:user_name/:space_name
 - /u/:user_name/:space_name/:tiddler_title

The :user_name option attempts to find the correct space/tiddler by
checking what both you and the user mentioned have access to