Invent "tiddlyimages" and add parser support
Basically, it's a normal tiddlylink to a tiddler but with an
exclamation mark before the link. It can be embedded inside a tiddlylink
title attribute or a transclusion (but is incompatible with embedding
inside markdown links).

The weirdness with the additions to the transclusion definition are
because parslet seems to hit a certain level of nesting before failing
and refuses to back out correctly. The fix in this case being to remove
the somewhat awkward unbracketed tiddlyimage inside transclusion
definition from the rest of the transclusion definition and create a new
one specifically for that.