Make markdown renderer use new link models
Note that if the markdown being rendered isn't linked to a tiddler (e.g.
the space description) then the old rendering method will be used. Also,
if a space/tiddler doesn't exist, then the shortlink will be used
instead of the full link.

This also changes the link format slightly so that both the tiddler and
the revision of the tiddler are linked to. This allows us to
(eventually) link revisions to the revision they were initially linked
to as well as to always link the latest version of the tiddler.

This provides a number of benefits:

1 - You don't need to parse tiddler markdown to extract links every time
you render it
2 - You can rename tiddlers as often as you like without breaking any
links
3 - You can create new tiddlers with the same title as often as you like
and still have the link pointing to the tiddler it initially pointed to
(this works even while editing the tiddler that contains the link as
long as the link itself doesn't change)
4 - If you then delete a tiddler, the link updates to the shortlink
format, meaning it still goes somewhere important